Rename Time file.
